Extracted component display to own component.
