+++ Frontend/course-flow-app/src/app/discovery/discovery-card-selectable/discovery-card-selectable.component.css
@@ -1,0 +1,26 @@
+.card-selectable-container-grid-container {
+    margin-top: 2em;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 2em;
+}
+
+.card-selectable-container-grid-item mat-card-content {
+    padding: 2em;
+    font-size: x-large;
+    font-weight: 500;
+    text-align: center;
+    min-height: 50px;
+    min-width: 200px;
+    align-self: center;
+}
+
+.card-selectable-container-grid-item mat-card-actions {
+    align-self: center;
+}
+
+.card-selectable-container h3 {
+    font-size: 20pt;
+    text-align: center;
+}
