+++ Frontend/course-flow-app/src/app/discovery/discovery-text-based-lists/discovery-text-based-lists.component.css
@@ -13,35 +13,3 @@
 .text-based-list-description {
     font-size: 15pt;
 }
-
-.text-based-list-back-button {
-    align-self: center;
-    width: 100px;
-}
-
-.text-based-list-container h3 {
-    font-size: 20pt;
-    text-align: center;
-}
-
-.text-based-list-grid-container {
-    margin-top: 2em;
-    display: grid;
-    grid-template-columns: 40% 40%;
-    justify-content: center;
-    gap: 2em;
-}
-
-.text-based-list-grid-item mat-card-content {
-    padding: 2em;
-    font-size: x-large;
-    font-weight: 500;
-    text-align: center;
-    min-height: 50px;
-    align-self: center;
-}
-
-.text-based-list-grid-item mat-card-actions {
-    align-self: center;
-}
-
