+++ Frontend/course-flow-app/src/app/discovery/discovery-text-based-lists/discovery-text-based-lists.component.html
@@ -1,14 +1,5 @@
 <div class="text-based-list-container">
 
-    <button 
-        class="text-based-list-back-button"
-        mat-raised-button color="primary" 
-        *ngIf="previousRootNode.length !== 0" 
-        (click)="onBackButtonPressed()"
-    > 
-        Back 
-    </button>
-
     <!-- The current "root" data/node that we are on will have information displayed in this section. -->
     <h2 class="text-based-list-title"> {{ currentRootNode.label }} </h2>
 
@@ -19,38 +10,8 @@
     <mat-divider/>
 
     <!-- All the connections will be listed here. -->
-    <div *ngFor="let connectionData of currentRootNode.outConnections;">
- 
-        <h3> {{connectionData.label}} </h3>
-
-        <div class="text-based-list-grid-container">
-
-            <div class="text-based-list-grid-item" >
- 
-                <div *ngFor="let connectionId of connectionData.connectionIds;">
-
-                    <mat-card>
-                        
-                        <mat-card-content>
-
-                            {{ getConnectionLabel(connectionId) }}
-
-                        </mat-card-content>
-
-                        <mat-card-actions>
-
-                            <button mat-raised-button color="primary" (click)="onViewMorePressed(connectionId)"> View More </button>
-        
-                        </mat-card-actions>
-
-                    </mat-card>
-
-                </div>
-
-            </div>
-
-        </div>
-
-    </div>
+    <app-discovery-card-selectable [connections]="currentRootNode.inConnections"/>
+    <app-discovery-card-selectable [connections]="currentRootNode.outConnections" />
+    <app-discovery-card-selectable [connections]="currentRootNode.coConnections"/>
 
 </div>
